+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Releases
 
+## v3.2.0
+
+* Adding in region config for backup services with `aws_backup_region_settings`
+* Adding in feature to allow cross-region backup copies
+* Add tagging to tagable resources
+
 ## v3.1.0
 
 * Support for Terraform versions 0.13 and above to (but not including) 1.0